ingredient auto moved to separate component
